generate structure for multiple UI tooling
