+++ src/common/player/mod.rs
@@ -1,4 +1,3 @@
-
 use rand::Rng;
 mod common;
